+++ integration-tests/http/src/utils/providers/tax-data-provider/index.ts
@@ -1,0 +1,8 @@
+import { ModuleProvider, Modules } from "@medusajs/framework/utils"
+import { TaxDataProviderService } from "./services/tax-data-provider"
+
+const services = [TaxDataProviderService]
+
+export default ModuleProvider(Modules.TAX, {
+  services,
+})