+++ packages/modules/providers/payment-stripe/src/core/stripe-base.ts
@@ -110,6 +110,10 @@ abstract class StripeBase extends AbstractPaymentProvider<StripeOptions> {
   ): Partial<Stripe.PaymentIntentCreateParams> {
     const res = {} as Partial<Stripe.PaymentIntentCreateParams>
 
+    const paymentMethodTypes =
+      (extra?.payment_method_types as string[] | undefined) ??
+      (this.paymentIntentOptions.payment_method_types as string[] | undefined)
+
     res.description = (extra?.payment_description ??
       this.options_?.paymentDescription) as string
 
@@ -122,9 +126,7 @@ abstract class StripeBase extends AbstractPaymentProvider<StripeOptions> {
       (extra?.setup_future_usage as "off_session" | "on_session" | undefined) ??
       this.paymentIntentOptions.setup_future_usage
 
-    res.payment_method_types =
-      (extra?.payment_method_types as string[]) ??
-      (this.paymentIntentOptions.payment_method_types as string[])
+    res.payment_method_types = paymentMethodTypes
 
     res.payment_method_data =
       extra?.payment_method_data as Stripe.PaymentIntentCreateParams.PaymentMethodData
@@ -147,6 +149,12 @@ abstract class StripeBase extends AbstractPaymentProvider<StripeOptions> {
 
     // @ts-expect-error - Need to update Stripe SDK
     res.shared_payment_token = extra?.shared_payment_token as string | undefined
+
+    if (!paymentMethodTypes?.length) {
+      res.payment_method_configuration =
+        (extra?.payment_method_configuration as string | undefined) ??
+        this.options_?.paymentMethodConfiguration
+    }
 
     return res
   }
@@ -702,14 +710,15 @@ abstract class StripeBase extends AbstractPaymentProvider<StripeOptions> {
         const paymentMethod =
           typeof webhookPaymentMethod === "object"
             ? (intent.payment_method as Stripe.PaymentMethod)
-            : await this.stripe_.paymentMethods.retrieve(
-                webhookPaymentMethod as string
-              )
+            : typeof webhookPaymentMethod === "string"
+            ? await this.stripe_.paymentMethods.retrieve(webhookPaymentMethod)
+            : null
 
         return {
-          action: this.isAsyncPaymentMethod(paymentMethod)
-            ? PaymentActions.PENDING_AUTHORIZATION
-            : PaymentActions.PENDING,
+          action:
+            paymentMethod && this.isAsyncPaymentMethod(paymentMethod)
+              ? PaymentActions.PENDING_AUTHORIZATION
+              : PaymentActions.PENDING,
           data: {
             session_id: intent.metadata.session_id,
             amount: getAmountFromSmallestUnit(intent.amount, currency),