+++ www/apps/user-guide/next.config.mjs
@@ -26,9 +26,6 @@ const withMDX = mdx({
         brokenLinkCheckerPlugin,
         {
           crossProjects: {
-            bloom: {
-              projectPath: path.resolve("..", "bloom"),
-            },
             docs: {
               projectPath: path.resolve("..", "book"),
             },
@@ -54,9 +51,6 @@ const withMDX = mdx({
         {
           baseUrl: process.env.NEXT_PUBLIC_BASE_URL,
           projectUrls: {
-            bloom: {
-              url: process.env.NEXT_PUBLIC_BLOOM_URL,
-            },
             docs: {
               url: process.env.NEXT_PUBLIC_DOCS_URL,
               path: "",