+++ www/packages/tags/src/tags/event.ts
@@ -2,9 +2,5 @@ export const event = [
   {
     "title": "Create Event Module",
     "path": "https://docs.medusajs.com/resources/infrastructure-modules/event/create"
-  },
-  {
-    "title": "Use Event Module",
-    "path": "https://docs.medusajs.com/resources/references/event-service"
   }
 ]